+++ commons/redis/redis.go
@@ -52,11 +52,22 @@ type RedisConnection struct {
 	lastRefreshInstant           time.Time
 	errLastSeen                  error
 	mu                           sync.RWMutex
+	PoolSize                     int
+	MinIdleConns                 int
+	ReadTimeout                  time.Duration
+	WriteTimeout                 time.Duration
+	DialTimeout                  time.Duration
+	PoolTimeout                  time.Duration
+	MaxRetries                   int
+	MinRetryBackoff              time.Duration
+	MaxRetryBackoff              time.Duration
 }
 
 // Connect initializes a Redis connection
 func (rc *RedisConnection) Connect(ctx context.Context) error {
 	rc.Logger.Info("Connecting to Redis/Valkey...")
+
+	rc.InitVariables()
 
 	var err error
 	if rc.UseGCPIAMAuth {
@@ -72,10 +83,19 @@ func (rc *RedisConnection) Connect(ctx context.Context) error {
 	}
 
 	opts := &redis.UniversalOptions{
-		Addrs:      rc.Address,
-		MasterName: rc.MasterName,
-		DB:         rc.DB,
-		Protocol:   rc.Protocol,
+		Addrs:           rc.Address,
+		MasterName:      rc.MasterName,
+		DB:              rc.DB,
+		Protocol:        rc.Protocol,
+		PoolSize:        rc.PoolSize,
+		MinIdleConns:    rc.MinIdleConns,
+		ReadTimeout:     rc.ReadTimeout,
+		WriteTimeout:    rc.WriteTimeout,
+		DialTimeout:     rc.DialTimeout,
+		PoolTimeout:     rc.PoolTimeout,
+		MaxRetries:      rc.MaxRetries,
+		MinRetryBackoff: rc.MinRetryBackoff,
+		MaxRetryBackoff: rc.MaxRetryBackoff,
 	}
 
 	if rc.UseGCPIAMAuth {
@@ -228,5 +248,44 @@ func (rc *RedisConnection) refreshTokenLoop(ctx context.Context) {
 		case <-ctx.Done():
 			return
 		}
+	}
+}
+
+// InitVariables sets default values for RedisConnection
+func (rc *RedisConnection) InitVariables() {
+	if rc.PoolSize == 0 {
+		rc.PoolSize = 10
+	}
+
+	if rc.MinIdleConns == 0 {
+		rc.MinIdleConns = 0
+	}
+
+	if rc.ReadTimeout == 0 {
+		rc.ReadTimeout = 3 * time.Second
+	}
+
+	if rc.WriteTimeout == 0 {
+		rc.WriteTimeout = 3 * time.Second
+	}
+
+	if rc.DialTimeout == 0 {
+		rc.DialTimeout = 5 * time.Second
+	}
+
+	if rc.PoolTimeout == 0 {
+		rc.PoolTimeout = 2 * time.Second
+	}
+
+	if rc.MaxRetries == 0 {
+		rc.MaxRetries = 3
+	}
+
+	if rc.MinRetryBackoff == 0 {
+		rc.MinRetryBackoff = 8 * time.Millisecond
+	}
+
+	if rc.MaxRetryBackoff == 0 {
+		rc.MaxRetryBackoff = 1 * time.Second
 	}
 }